+++ src/Oro/Bundle/ThemeBundle/Tests/Functional/DataFixtures/LoadThemeConfigurationData.php
@@ -5,7 +5,6 @@ namespace Oro\Bundle\ThemeBundle\Tests\Functional\DataFixtures;
 use Doctrine\Common\DataFixtures\AbstractFixture;
 use Doctrine\ORM\EntityManager;
 use Doctrine\Persistence\ObjectManager;
-use Oro\Bundle\ThemeBundle\Entity\Enum\ThemeConfigurationType;
 use Oro\Bundle\ThemeBundle\Entity\ThemeConfiguration;
 use Oro\Bundle\UserBundle\DataFixtures\UserUtilityTrait;
 
@@ -13,18 +12,18 @@ class LoadThemeConfigurationData extends AbstractFixture
 {
     use UserUtilityTrait;
 
-    public const THEME_CONFIGURATION_1 = 'Theme 1';
+    public const string THEME_CONFIGURATION_1 = 'Theme 1';
 
     protected array $themeConfigurationsData = [
         [
             'name' => self::THEME_CONFIGURATION_1,
-            'type' => ThemeConfigurationType::Storefront,
+            'type' => 'storefront',
             'theme' => 'default',
         ],
     ];
 
     #[\Override]
-    public function load(ObjectManager $manager)
+    public function load(ObjectManager $manager): void
     {
         /** @var EntityManager $manager */
         $user = $this->getFirstUser($manager);
@@ -34,7 +33,7 @@ class LoadThemeConfigurationData extends AbstractFixture
         foreach ($this->themeConfigurationsData as $themeConfigurationData) {
             $themeConfiguration = (new ThemeConfiguration())
                 ->setName($themeConfigurationData['name'])
-                ->setType($themeConfigurationData['type'] ?? ThemeConfigurationType::Storefront)
+                ->setType($themeConfigurationData['type'])
                 ->setConfiguration($this->processConfiguration($themeConfigurationData['configuration'] ?? []))
                 ->setTheme($themeConfigurationData['theme'])
                 ->setOwner($businessUnit)